Improved server code clarity (edk-route-verifier-114) Removed language and debug switches from verify request Added npm scripts for all flavours of verifier
