+++ package.json
@@ -25,6 +25,8 @@
     "publish:patch": "sh scripts/publish-version.sh patch",
     "publish:prerelease": "sh scripts/publish-version.sh prerelease",
     "server": "node --experimental-modules --experimental-json-modules src/cli/cli.js server -c conf/config.json",
+    "browser": "node --experimental-modules --experimental-json-modules src/cli/cli.js browser -c conf/config.json",
+    "file": "node --experimental-modules --experimental-json-modules src/cli/cli.js file -c conf/config.json",
     "test": "npm run lint && npm run apiTest"
   },
   "dependencies": {
