+++ src/server/pages/index.ejs
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="<%= language %>">
+<html lang="pl">
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -45,19 +45,6 @@
     <aside class="main-sidebar">
         <!-- sidebar: style can be found in sidebar.less -->
         <section class="sidebar">
-            <div class="box-body bg-gray">
-                <h4>
-                    Język błędów
-                </h4>
-                <% for(var i = 0; i < languages.length; i++) { %>
-                    <a href="?lang=<%= languages[i] %>">
-                        <button id="changeLanguage" type="button" class="btn btn-app <%= language === languages[i] ? 'bg-green' : '' %>">
-                            <i class="fa fa-language"></i> <%= languages[i].toUpperCase() %>
-                        </button>
-                    </a>
-                <% } %>
-            </div>
-
             <div class="box-body bg-gray">
                 <h4>
                     Trasy
