+++ src/server/server.js
@@ -1,34 +1,13 @@
-import browserify from 'browserify';
-import basicAuth from 'express-basic-auth';
 import bodyParser from 'body-parser';
 import cors from 'cors';
-import helmet from 'helmet';
 import express from 'express';
-import fs from 'fs';
 import logger from 'loglevel';
-import path from 'path';
 
+import { addBrowserRoutes } from './browserRoutes.js';
+import { secureServer } from './security.js';
 import verifyRoute from '../core/verifyRoute.js';
 import RouteVerificationInput from '../data/RouteVerificationInput.js';
 import RouteVerificationOptions from '../data/RouteVerificationOptions.js';
-
-function secureServer(app, config) {
-    app.use(helmet());
-
-    const { apiUser, apiPass } = config;
-    if (apiUser && apiPass) {
-        app.use(
-            basicAuth({
-                users: { [apiUser]: apiPass }
-            })
-        );
-    } else {
-        logger.warn(
-            'Server API username/password not defined in the configuration. ' +
-                'Starting server without Basic Authentication.'
-        );
-    }
-}
 
 function setupLogger(debug) {
     if (debug) {
@@ -38,76 +17,21 @@ function setupLogger(debug) {
     }
 }
 
-function addStaticRoutes(app, config, port) {
-    const languagesPath = path.resolve('src/core/lang');
-    const languages = fs
-        .readdirSync(languagesPath)
-        .filter(file => file.endsWith('.json'))
-        .map(file => file.replace('.json', ''));
-    const resources = [];
-    fs.readdirSync(path.resolve(config.resourcesPath)).forEach(file => {
-        if (file.search(/\.kml$/i) >= 0) {
-            resources.push(file.replace('.kml', ''));
-        }
-    });
-
-    // set the view engine to ejs
-    app.set('views', path.resolve('src/server'));
-    app.set('view engine', 'ejs');
-    app.use('/pages', express.static('src/server/pages'));
-    app.use('/browser/loadMap.js', express.static('src/browser/loadMap.js'));
-    app.get('/browser/verifier.js', (req, res) => {
-        res.setHeader('Content-Type', 'application/javascript');
-        browserify(path.resolve('src/browser/verifier.js'))
-            .transform('babelify', { presets: ['@babel/preset-env'] })
-            .bundle()
-            .pipe(res);
-    });
-
-    // all resources page
-    app.get('/resources', (req, res) => {
-        res.render('pages/resources', {
-            resources
-        });
-    });
-
-    app.get('/', (req, res) => {
-        res.redirect('/resources');
-    });
-
-    // index page
-    app.get('/:routeId', (req, res) => {
-        const id = req.params.routeId;
-        const { lang = 'pl' } = req.query;
-        res.render('pages/index', {
-            googleMapsApiKey: config.googleMapsApiKey,
-            routeId: id,
-            serverPort: port,
-            language: lang,
-            languages,
-            resources
-        });
-    });
-
-    app.get('/kml/:routeId', cors(), (req, res) => {
-        const id = req.params.routeId;
-        logger.info(`Sending KML for route ${id}.`);
-        res.sendFile(path.resolve(path.join(config.resourcesPath, `${id}.kml`)));
-    });
-}
-
-// eslint-disable-next-line import/prefer-default-export
 export function startServer(config, port = 9102, language = 'en', debug = false, serveWebContent = false) {
     const app = express();
 
-    secureServer(app, config);
+    if (serveWebContent) {
+        addBrowserRoutes(app, config, port);
+    } else {
+        secureServer(app, config);
+    }
     setupLogger(debug);
 
     app.post('/api/verify', cors(), bodyParser.json(), (req, res) => {
-        const { kml, language: l = language, debug: d = debug } = req.body;
+        const { kml } = req.body;
 
         const routeData = new RouteVerificationInput(kml);
-        const verificationOption = new RouteVerificationOptions(config, l, d);
+        const verificationOption = new RouteVerificationOptions(config, language, debug);
 
         verifyRoute(routeData, verificationOption)
             .then(verificationOutput => res.send(verificationOutput))
@@ -120,12 +44,13 @@ export function startServer(config, port = 9102, language = 'en', debug = false,
             });
     });
 
-    if (serveWebContent) {
-        addStaticRoutes(app, config, port);
-        logger.info(`Open browser at http://localhost:${port}.`);
-    } else {
-        logger.info(`Starting Verify API server at http://localhost:${port}.`);
-    }
+    logger.info(`Starting Verify API server...`);
 
-    app.listen(port);
+    app.listen(port, () => {
+        logger.info(
+            `Started serving Verify API ${serveWebContent ? 'and static content ' : ''}at http://localhost:${port}.`
+        );
+    });
 }
+
+export default startServer;
